Move remainder of GenericSourceTests to MethodSourceTests
