+++ platform-tests/src/test/java/org/junit/platform/engine/support/descriptor/MethodSourceTests.java
@@ -10,13 +10,17 @@
 
 package org.junit.platform.engine.support.descriptor;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import java.io.Serializable;
 import java.lang.reflect.Method;
+import java.util.stream.Stream;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInfo;
 import org.junit.platform.commons.util.PreconditionViolationException;
 
 /**
@@ -24,7 +28,32 @@ import org.junit.platform.commons.util.PreconditionViolationException;
  *
  * @since 1.0
  */
-class MethodSourceTests {
+class MethodSourceTests extends AbstractTestSourceTests {
+
+	@Override
+	Stream<Serializable> createSerializableInstances() throws Exception {
+		return Stream.of( //
+			MethodSource.from(getMethod("method1")), //
+			MethodSource.from(getMethod("method2")) //
+		);
+	}
+
+	@Test
+	void methodSource() throws Exception {
+		Method testMethod = getMethod("method1");
+		MethodSource source = MethodSource.from(testMethod);
+
+		assertThat(source.getClassName()).isEqualTo(getClass().getName());
+		assertThat(source.getMethodName()).isEqualTo(testMethod.getName());
+		assertThat(source.getMethodParameterTypes()).isEqualTo(String.class.getName());
+	}
+
+	@Test
+	void equalsAndHashCodeForMethodSource(TestInfo testInfo) throws Exception {
+		Method method1 = getMethod("method1");
+		Method method2 = getMethod("method2");
+		assertEqualsAndHashCode(MethodSource.from(method1), MethodSource.from(method1), MethodSource.from(method2));
+	}
 
 	@Test
 	void instantiatingWithNullNamesShouldThrowPreconditionViolationException() {
@@ -167,6 +196,16 @@ class MethodSourceTests {
 		Method m2 = Byte.class.getDeclaredMethod("byteValue");
 
 		assertNotEquals(MethodSource.from(m1).hashCode(), MethodSource.from(m2).hashCode());
+	}
+
+	private Method getMethod(String name) throws Exception {
+		return getClass().getDeclaredMethod(name, String.class);
+	}
+
+	void method1(String text) {
+	}
+
+	void method2(String text) {
 	}
 
 }
